+++ bridges/quickjs/js/src/embedded-react/index.js
@@ -48,6 +48,7 @@ export {StyleSheet} from './StyleSheet.js';
 export {Platform} from './Platform.js';
 export {AppRegistry} from './AppRegistry.js';
 export {Animated, useAnimatedValue} from './Animated.js';
+export {PanResponder} from './PanResponder.js';
 export {usePersistentState} from './usePersistentState.js';
 export {useHostValue} from './useHostValue.js';
 export {Easing} from './Easing.js';